chore: move NitpickExtension.run() WIP
